Re-generated with latest STM32CubeMX - also added Timer 4
Timer 4 is hoped to be used for the GPS timestamp where it is reset to zero each 1pps pulse  from the GPS (unlikn Timer 3 at present) No functionality changes yet... Timer 3 will become redundant if this can be made to work properly.

It means using PD13 for the 1pps pulse instead of PB10.